debugging and adding the AI to the UI

+++ src/main/java/com/example/trainerintake/model/Question.java
@@ -8,14 +8,14 @@ public class Question {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "question_id")
-    private Integer questionId;
+    private Long questionId;
 
     @Column(name = "prompt")  // ✅ this is the actual question text
     private String questionText;
 
     // Getters and setters
-    public Integer getQuestionId() { return questionId; }
-    public void setQuestionId(Integer questionId) { this.questionId = questionId; }
+    public Long getQuestionId() { return questionId; }
+    public void setQuestionId(Long questionId) { this.questionId = questionId; }
 
     public String getQuestionText() { return questionText; }
     public void setQuestionText(String questionText) { this.questionText = questionText; }
